# internal/transport/http/docs/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Task Service API",
    "version": "2.0.0",
    "description": "API for task templates and concrete task occurrences."
  },
  "servers": [
    {
      "url": "http://localhost:8080"
    }
  ],
  "tags": [
    {
      "name": "Health",
      "description": "Service health"
    },
    {
      "name": "Tasks",
      "description": "Concrete task occurrences"
    },
    {
      "name": "Task Templates",
      "description": "Recurring task templates"
    }
  ],
  "paths": {
    "/healthz": {
      "get": {
        "tags": ["Health"],
        "summary": "Health check",
        "responses": {
          "200": {
            "description": "Service is healthy",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Health"
                }
              }
            }
          },
          "503": {
            "$ref": "#/components/responses/ServiceUnavailable"
          }
        }
      }
    },
    "/api/v1/tasks": {
      "get": {
        "tags": ["Tasks"],
        "summary": "List task occurrences",
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 100,
              "default": 50
            }
          },
          {
            "name": "offset",
            "in": "query",
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0
            }
          },
          {
            "name": "status",
            "in": "query",
            "schema": {
              "type": "string",
              "enum": ["new", "in_progress", "done"]
            }
          },
          {
            "name": "template_id",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int64",
              "minimum": 1
            }
          },
          {
            "name": "date_from",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "date"
            }
          },
          {
            "name": "date_to",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "date"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Task occurrence list. Pagination metadata is returned in X-Total-Count, X-Limit and X-Offset headers.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Task"
                  }
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          }
        }
      },
      "post": {
        "tags": ["Tasks"],
        "summary": "Create a task occurrence entry point",
        "description": "Creates a task template and returns the first generated occurrence. For recurring schedules, future occurrences are generated lazily.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateTaskRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Created task occurrence",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Task"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          }
        }
      }
    },
    "/api/v1/tasks/{id}": {
      "get": {
        "tags": ["Tasks"],
        "summary": "Get task occurrence by ID",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64",
              "minimum": 1
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Task occurrence",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Task"
                }
              }
            }
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          }
        }
      },
      "put": {
        "tags": ["Tasks"],
        "summary": "Update a task occurrence",
        "description": "Updates a concrete task occurrence. Recurring schedule changes must be done through task template endpoints.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64",
              "minimum": 1
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateTaskRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Updated task occurrence",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Task"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          }
        }
      },
      "delete": {
        "tags": ["Tasks"],
        "summary": "Delete a task occurrence",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64",
              "minimum": 1
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Deleted"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          }
        }
      }
    },
    "/api/v1/task-templates": {
      "get": {
        "tags": ["Task Templates"],
        "summary": "List task templates",
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 100,
              "default": 50
            }
          },
          {
            "name": "offset",
            "in": "query",
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0
            }
          },
          {
            "name": "search",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "schedule_type",
            "in": "query",
            "schema": {
              "type": "string",
              "enum": ["daily", "monthly", "specific_dates", "odd_days", "even_days"]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Task template list. Pagination metadata is returned in X-Total-Count, X-Limit and X-Offset headers.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/TaskTemplate"
                  }
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          }
        }
      },
      "post": {
        "tags": ["Task Templates"],
        "summary": "Create a task template",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateTaskTemplateRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Created task template",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TaskTemplate"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          }
        }
      }
    },
    "/api/v1/task-templates/{id}": {
      "get": {
        "tags": ["Task Templates"],
        "summary": "Get task template by ID",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64",
              "minimum": 1
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Task template",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TaskTemplate"
                }
              }
            }
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          }
        }
      },
      "put": {
        "tags": ["Task Templates"],
        "summary": "Update task template",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64",
              "minimum": 1
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateTaskTemplateRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Updated task template",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TaskTemplate"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          }
        }
      },
      "delete": {
        "tags": ["Task Templates"],
        "summary": "Delete task template and all its occurrences",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64",
              "minimum": 1
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Deleted"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          }
        }
      }
    }
  },
  "components": {
    "responses": {
      "BadRequest": {
        "description": "Bad request",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      },
      "NotFound": {
        "description": "Resource not found",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      },
      "ServiceUnavailable": {
        "description": "Service is unhealthy",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      }
    },
    "schemas": {
      "CreateTaskRequest": {
        "type": "object",
        "required": ["title"],
        "properties": {
          "title": {
            "type": "string",
            "example": "Daily patient calls"
          },
          "description": {
            "type": "string",
            "example": "Follow-up after discharge"
          },
          "status": {
            "type": "string",
            "enum": ["new", "in_progress", "done"],
            "example": "new"
          },
          "schedule": {
            "$ref": "#/components/schemas/Schedule"
          }
        }
      },
      "UpdateTaskRequest": {
        "type": "object",
        "required": ["title", "status"],
        "properties": {
          "title": {
            "type": "string",
            "example": "Daily patient calls"
          },
          "description": {
            "type": "string",
            "example": "Follow-up after discharge"
          },
          "status": {
            "type": "string",
            "enum": ["new", "in_progress", "done"],
            "example": "done"
          },
          "schedule": {
            "$ref": "#/components/schemas/Schedule"
          }
        }
      },
      "CreateTaskTemplateRequest": {
        "type": "object",
        "required": ["title"],
        "properties": {
          "title": {
            "type": "string",
            "example": "Inventory"
          },
          "description": {
            "type": "string",
            "example": "Monthly inventory check"
          },
          "schedule": {
            "$ref": "#/components/schemas/Schedule"
          }
        }
      },
      "UpdateTaskTemplateRequest": {
        "type": "object",
        "required": ["title"],
        "properties": {
          "title": {
            "type": "string",
            "example": "Inventory"
          },
          "description": {
            "type": "string",
            "example": "Monthly inventory check"
          },
          "schedule": {
            "$ref": "#/components/schemas/Schedule"
          }
        }
      },
      "Schedule": {
        "type": "object",
        "required": ["type"],
        "properties": {
          "type": {
            "type": "string",
            "enum": ["daily", "monthly", "specific_dates", "odd_days", "even_days"]
          },
          "every_days": {
            "type": "integer",
            "minimum": 1,
            "example": 2
          },
          "day_of_month": {
            "type": "integer",
            "minimum": 1,
            "maximum": 30,
            "example": 15
          },
          "dates": {
            "type": "array",
            "items": {
              "type": "string",
              "format": "date"
            },
            "example": ["2026-04-10", "2026-04-15"]
          },
          "start_date": {
            "type": "string",
            "format": "date",
            "example": "2026-04-08"
          },
          "timezone": {
            "type": "string",
            "description": "IANA timezone identifier. Defaults to UTC.",
            "example": "Europe/Moscow"
          }
        }
      },
      "Task": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64",
            "example": 101
          },
          "template_id": {
            "type": "integer",
            "format": "int64",
            "example": 10
          },
          "title": {
            "type": "string",
            "example": "Daily patient calls"
          },
          "description": {
            "type": "string",
            "example": "Follow-up after discharge"
          },
          "status": {
            "type": "string",
            "enum": ["new", "in_progress", "done"],
            "example": "new"
          },
          "scheduled_for": {
            "type": "string",
            "format": "date",
            "example": "2026-04-08"
          },
          "schedule": {
            "$ref": "#/components/schemas/Schedule"
          },
          "completed_at": {
            "type": "string",
            "format": "date-time",
            "nullable": true,
            "example": "2026-04-08T10:30:00Z"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "example": "2026-04-08T09:00:00Z"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "example": "2026-04-08T09:00:00Z"
          }
        }
      },
      "TaskTemplate": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64",
            "example": 10
          },
          "title": {
            "type": "string",
            "example": "Inventory"
          },
          "description": {
            "type": "string",
            "example": "Monthly inventory check"
          },
          "schedule": {
            "$ref": "#/components/schemas/Schedule"
          },
          "next_occurrence_on": {
            "type": "string",
            "format": "date",
            "nullable": true,
            "example": "2026-04-15"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "example": "2026-04-08T09:00:00Z"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "example": "2026-04-08T09:00:00Z"
          }
        }
      },
      "Health": {
        "type": "object",
        "properties": {
          "status": {
            "type": "string",
            "example": "ok"
          }
        }
      },
      "Error": {
        "type": "object",
        "properties": {
          "error": {
            "type": "string",
            "example": "invalid task input: limit must be between 1 and 100"
          }
        }
      }
    }
  }
}